requirements.txt
Django>=5.2.8,<5.3
djangorestframework>=3.16,<3.17
django-cors-headers>=4.3,<5
whitenoise>=6.6,<7
dj-database-url>=2.1,<3
psycopg[binary]>=3.1,<4
Pillow>=9.0
python-dotenv==1.0.0
gunicorn==21.0.0
